- hot fix - authentication system timer and maximum tries added

+++ porsline_config/user_app/serializers.py
@@ -1,4 +1,6 @@
 import re
+
+from django.utils import timezone
 from rest_framework import serializers
 from django.contrib.auth import get_user_model
 from rest_framework_simplejwt.settings import api_settings
@@ -7,6 +9,7 @@ from question_app.models import Folder
 from question_app.question_app_serializers import general_serializers
 from .tasks import send_otp
 from .models import OTPToken
+from porsline_config import settings
 
 
 class FolderSerializer(serializers.ModelSerializer):
@@ -52,11 +55,15 @@ class GateWaySerializer(serializers.Serializer):
 
 class OTPCheckSerializer(serializers.Serializer):
     token = serializers.CharField(max_length=5, min_length=5, required=True)
+    phone_number = serializers.CharField(max_length=20, min_length=11, required=True)
     refresh = serializers.CharField(read_only=True)
     access = serializers.CharField(read_only=True)
 
     def validate(self, data):
         token = data.get('token')
+        phone_number = data.get('phone_number')
+        if not re.match(r'^09\d{9}$', phone_number):
+            raise serializers.ValidationError("فرمت شماره تلفن صحیح نمی‌باشد.")
         if token is None:
             raise serializers.ValidationError("کد فعال‌سازی را وارد کنید.")
         if len(token) != 5:
@@ -64,16 +71,30 @@ class OTPCheckSerializer(serializers.Serializer):
         return data
 
     def create(self, validated_data):
+        phone_number = validated_data.get('phone_number')
         token = validated_data.get('token')
-        otp = OTPToken.objects.filter(token=token)
+        otp = OTPToken.objects.filter(user__phone_number=phone_number)
         if otp.exists():
             otp = otp.first()
+            otp.try_count += 1
+            otp.save()
+            if otp.try_count > settings.OTP_TRY_COUNT:
+                otp.delete()
+                raise serializers.ValidationError(
+                    "تعداد تلاش‌ هابیش از حد مجاز است، لطفا مجددا برای دریافت کد اقدام کنید.")
+            if otp.token != token:
+                raise serializers.ValidationError("کد فعال سازی منقضی شده یا اشتباه است.")
+            if timezone.now() - otp.created_at > timezone.timedelta(minutes=settings.OTP_LIFE_TIME):
+                otp.delete()
+                raise serializers.ValidationError("کد فعال سازی منقضی شده یا اشتباه است.")
             user = otp.user
             access_token = AccessToken.for_user(user)
             refresh_token = RefreshToken.for_user(user)
             otp.delete()
-            return {'token': token, 'refresh': str(refresh_token), 'access': str(access_token)}
-        raise serializers.ValidationError("کد فعال‌سازی صحیح نمی‌باشد.")
+            return {'token': otp.token, 'phone_number': otp.user.phone_number, 'refresh': str(refresh_token),
+                    'access': str(access_token)}
+        raise serializers.ValidationError(
+            "شماره تلفن یا کد فعال سازی صحیح نمی باشد لطفا مجددا برای دریافت کد اقدام کنید.")
 
 
 class RefreshTokenSerializer(serializers.Serializer):
